# clients/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OMR Engine API",
    "description": "Read OMR sheets (bubbles, barcodes, QR codes, OCR, ICR), manage templates, run bulk jobs and review uncertain reads.",
    "version": "1.0"
  },
  "paths": {
...
    "/tools/colors": {
      "post": {
        "tags": [
          "tools"
        ],
        "summary": "Sheet Colors",
        "description": "Main colours of a sheet with dropout suggestions for print colours.",
        "operationId": "sheet_colors_tools_colors_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_sheet_colors_tools_colors_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tools/dropout-suggest": {
      "post": {
        "tags": [
          "tools"
        ],
        "summary": "Dropout Suggest",
        "description": "Dropout settings that remove `target` while keeping `keep` dark.",
        "operationId": "dropout_suggest_tools_dropout_suggest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tools/dropout-preview": {
      "post": {
        "tags": [
          "tools"
        ],
        "summary": "Dropout Preview",
        "description": "The grey image the reader would see with these dropout settings (PNG).",
        "operationId": "dropout_preview_tools_dropout_preview_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_dropout_preview_tools_dropout_preview_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_dropout_preview_tools_dropout_preview_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "settings": {
            "type": "string",
            "title": "Settings",
            "description": "colorDropout JSON, e.g. {\"mode\": \"red\"} or \"grey\"",
            "default": "{}"
          },
          "max_width": {
            "type": "integer",
            "maximum": 1600.0,
            "minimum": 64.0,
            "title": "Max Width",
            "default": 1000
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_dropout_preview_tools_dropout_preview_post"
      },
...
      "Body_sheet_colors_tools_colors_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "A sample sheet (image or PDF)"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          },
          "k": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 2.0,
            "title": "K",
            "default": 6
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_sheet_colors_tools_colors_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SuggestBody": {
        "properties": {
          "target": {
            "type": "string",
            "title": "Target",
            "description": "Colour to remove, '#RRGGBB'"
          },
          "keep": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keep",
            "description": "Colours that must stay dark (default: black, pencil, blue pens)"
          }
        },
        "type": "object",
        "required": [
          "target"
        ],
        "title": "SuggestBody"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}